adds RepositoryFile and renames to RemoteRepositories impl class
